conf: Move `Toml` to its own module
